Move `get_remote_for_branch` to `BranchesMixin`
